Refactor: Refactoring code for ban-core-api
