+++ apps/ban-core-api/src/routes/data/diff/diff.ts
@@ -19,7 +19,7 @@ import {
   getDiffObjLine,
   closeCursor,
   streamCursorData,
-} from '../helper.js'
+} from '../helpers/index.js'
 
 const FETCH_SIZE = 500;
