+++ apps/ban-core-api/src/routes/data/helpers/config.ts
@@ -1,0 +1,2 @@
+export const VALID_FORMATS = ['raw', 'ban', 'standard-fr', 'standard-fr-int'];
+export const VALID_DATA_TYPES = ["district", "toponym", "address"];
